+++ src/jsonData/upcoming-events.js
@@ -1,0 +1,48 @@
+import profileImg from '../svgs/profile-icon.svg';
+
+const events = [
+    {
+      _id: 1,
+      imgSrc: profileImg,
+      createdBy: 'Julian Richards',
+      description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
+      tag: '+1 RSVP'
+    },
+    {
+      _id: 2,
+      imgSrc: profileImg,
+      createdBy: 'Julian Richards',
+      description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
+      tag: ''
+    },
+    {
+      _id: 3,
+      imgSrc: profileImg,
+      createdBy: 'Julian Richards',
+      description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
+      tag: '+1 RSVP'
+    },
+    {
+      _id: 4,
+      imgSrc: profileImg,
+      createdBy: 'Julian Richards',
+      description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
+      tag: ''
+    },
+    {
+      _id: 5,
+      imgSrc: profileImg,
+      createdBy: 'Julian Richards',
+      description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
+      tag: ''
+    },
+    {
+      _id: 6,
+      imgSrc: profileImg,
+      createdBy: 'Julian Richards',
+      description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
+      tag: ''
+    }
+]
+
+export default events;
