+++ src/user/dashboard/news-feed/news-feed.js
@@ -1,13 +1,10 @@
 import React, { Component} from "react";
 import { Button , Dropdown} from "react-bootstrap";
 import "./news-feed.scss";
-import profileImg from "../../../svgs/evt-creator.svg";
 import upVoteImg from "../../../svgs/up.svg";
 import downVoteImg from "../../../svgs/down.svg";
 import commentIcon from "../../../svgs/comment.svg";
-import eventImg from "../../../svgs/event-img-1.svg";
-import eventImg2 from "../../../svgs/event-img-2.svg";
-
+import feed from "../../../jsonData/news-feed";
 
 const CustomToggle = React.forwardRef(({ children, onClick }, ref) => (
   <a
@@ -39,88 +36,6 @@ const CustomMenu = React.forwardRef(
   },
 );
 
-const feed = [
-  {
-    _id: 1,
-    imgSrc: profileImg,
-    createdBy: 'Marjorie Alexander',
-    created: 'June 5, 2018 4:31 AM', 
-    details: 'ex sit ex laboris adipisicing enim eiusmod proident exercitation ea fugiat in mollit pariatur occaecat ut nostrud ullamco ex official',
-    upvotes: 810,
-    downVotes: 185,
-    eventImage: null,
-    type: 'Donut'
-  },
-  {
-    _id: 2,
-    imgSrc: profileImg,
-    createdBy: 'Marjorie Alexander',
-    created: 'June 5, 2018 4:31 AM',
-    details: 'ex sit ex laboris adipisicing enim eiusmod proident exercitation ea fugiat in mollit pariatur occaecat ut nostrud ullamco ex official',
-    upvotes: 80,
-    downVotes: 15,
-    eventImage: eventImg2,
-    type: 'Project',
-    projectName: 'Some Huge Project Name',
-    projectOwner: 'Neilson'
-  },
-  {
-    _id: 3,
-    imgSrc: profileImg,
-    createdBy: 'Marjorie Alexander',
-    created: 'June 5, 2018 4:31 AM',
-    details: 'ex sit ex laboris adipisicing enim eiusmod proident exercitation ea fugiat in mollit pariatur occaecat ut nostrud ullamco ex official',
-    upvotes: 104,
-    downVotes: 25,
-    eventImage: null,
-    type: 'Donut'
-  },
-  {
-    _id: 4,
-    imgSrc: profileImg,
-    createdBy: 'Marjorie Alexander',
-    created: 'June 5, 2018 4:31 AM',
-    details: 'ex sit ex laboris adipisicing enim eiusmod proident exercitation ea fugiat in mollit pariatur occaecat ut nostrud ullamco ex official',
-    upvotes: 810,
-    downVotes: 185,
-    eventImage: null,
-    type: 'Donut'
-  },
-  {
-    _id: 5,
-    imgSrc: profileImg,
-    createdBy: 'Marjorie Alexander',
-    created: 'June 5, 2018 4:31 AM',
-    eventName: 'Christmas Party',
-    details: 'ex sit ex laboris adipisicing enim eiusmod proident exercitation ea fugiat in mollit pariatur occaecat ut nostrud ullamco ex official',
-    upvotes: 80,
-    downVotes: 15,
-    eventImage: eventImg,
-    type: 'Event'
-  },
-  {
-    _id: 6,
-    imgSrc: profileImg,
-    createdBy: 'Marjorie Alexander',
-    created: 'June 5, 2018 4:31 AM',
-    details: 'ex sit ex laboris adipisicing enim eiusmod proident exercitation ea fugiat in mollit pariatur occaecat ut nostrud ullamco ex official',
-    upvotes: 80,
-    downVotes: 15,
-    eventImage: null,
-    type: 'Donut'
-  },
-  {
-    _id: 7,
-    imgSrc: profileImg,
-    createdBy: 'Marjorie Alexander',
-    created: 'June 5, 2018 4:31 AM',
-    details: 'ex sit ex laboris adipisicing enim eiusmod proident exercitation ea fugiat in mollit pariatur occaecat ut nostrud ullamco ex official',
-    upvotes: 810,
-    downVotes: 185,
-    eventImage: null,
-    type: 'Donut'
-  },
-]
 
 class NewsFeed extends Component {
   state = { date: new Date() };
