+++ src/user/dashboard/notifications/notifications.js
@@ -1,39 +1,6 @@
 import React, { Component } from "react";
-import donutIcon from '../../../svgs/donut-icon.svg'
-import icon2 from '../../../svgs/not-icon-2.svg'
-import icon3 from '../../../svgs/not-icon-3.svg'
 import "./notifications.scss";
-
-const customNotifications = [
-  {
-    _id: 1,
-    imgSrc: donutIcon,
-    heading: 'Cody Nguyen',
-    content: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: '+1 RSVP'
-  },
-  {
-    _id: 2,
-    imgSrc: icon2,
-    heading: 'Arlene Mccoy',
-    content: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: ''
-  },
-  {
-    _id: 3,
-    imgSrc: icon3,
-    heading: 'Julian Richards',
-    content: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: 'View'
-  },
-  {
-    _id: 4,
-    imgSrc: donutIcon,
-    heading: 'Julian Richards',
-    content: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: ''
-  },
-]
+import customNotifications from '../../../jsonData/notifications'
 
 class Notifications extends Component {
   render() {
