+++ src/user/dashboard/upcoming-events/upcoming-events.js
@@ -1,51 +1,6 @@
 import React, { Component } from "react";
 import "./upcoming-events.scss";
-import profileImg from '../../../svgs/profile-icon.svg';
-
-const events = [
-  {
-    _id: 1,
-    imgSrc: profileImg,
-    createdBy: 'Julian Richards',
-    description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: '+1 RSVP'
-  },
-  {
-    _id: 2,
-    imgSrc: profileImg,
-    createdBy: 'Julian Richards',
-    description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: ''
-  },
-  {
-    _id: 3,
-    imgSrc: profileImg,
-    createdBy: 'Julian Richards',
-    description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: '+1 RSVP'
-  },
-  {
-    _id: 4,
-    imgSrc: profileImg,
-    createdBy: 'Julian Richards',
-    description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: ''
-  },
-  {
-    _id: 5,
-    imgSrc: profileImg,
-    createdBy: 'Julian Richards',
-    description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: ''
-  },
-  {
-    _id: 6,
-    imgSrc: profileImg,
-    createdBy: 'Julian Richards',
-    description: 'Lorem ipsum dolor sit amet, duo esse augue torgatos te, ius an nisi deterruisset.',
-    tag: ''
-  }
-]
+import events from '../../../jsonData/upcoming-events';
 
 class UpcomingEvents extends Component {
   render() {
